test: move range-align into new-range and fix tests exceed time problem
